+++ package.json
@@ -20,9 +20,8 @@
   },
   "scripts": {
     "clean:node_modules": "rm -rf node_modules/**/node_modules && rm -rf node_modules/@types/**/node_modules && rm -rf node_modules/@augu/**/node_modules",
-    "docker:build:edge": "docker buuild . -t auguwu/nino:edge",
+    "docker:build:edge": "docker build . -t auguwu/nino:edge",
     "clean:win:tar": "cp node_modules/@augu/collections/build/index.js.* node_modules/@augu/collections/build/index.js && rm node_modules/@augu/collections/build/index.js.*",
-    "husky:install": "husky install",
     "docker:build": "docker build . -t auguwu/nino:latest",
     "migrations": "npm run build && typeorm migration:run && typeorm schema:sync",
     "shortlinks": "node scripts/shortlinks.js",
@@ -37,13 +36,13 @@
     "@augu/collections": "1.0.5",
     "@augu/dotenv": "1.3.0",
     "@augu/http": "1.2.0",
-    "@augu/lilith": "1.2.0",
+    "@augu/lilith": "2.2.0",
     "@augu/orchid": "2.2.3",
-    "@augu/utils": "1.3.0",
-    "@kubernetes/client-node": "0.14.0",
-    "@sentry/node": "6.2.3",
+    "@augu/utils": "1.4.2",
+    "@kubernetes/client-node": "0.14.1",
+    "@sentry/node": "6.2.5",
     "eris": "github:abalabahaha/eris#dev",
-    "ioredis": "4.24.5",
+    "ioredis": "4.25.0",
     "js-yaml": "4.0.0",
     "ms": "2.1.3",
     "pg": "8.5.1",
@@ -51,20 +50,20 @@
     "reflect-metadata": "0.1.13",
     "source-map-support": "0.5.19",
     "tslog": "3.1.2",
-    "typeorm": "0.2.31",
+    "typeorm": "0.2.32",
     "ws": "7.4.4"
   },
   "devDependencies": {
     "@augu/eslint-config": "1.10.0",
     "@augu/tsconfig": "1.0.1",
-    "@types/common-tags": "1.8.0",
-    "@types/ioredis": "4.22.0",
-    "@types/js-yaml": "3.12.4",
+    "@types/express": "4.17.11",
+    "@types/ioredis": "4.22.2",
+    "@types/js-yaml": "4.0.0",
     "@types/ms": "0.7.31",
     "@types/node": "14.14.37",
-    "@types/ws": "7.4.0",
-    "@typescript-eslint/eslint-plugin": "4.19.0",
-    "@typescript-eslint/parser": "4.19.0",
+    "@types/ws": "7.4.1",
+    "@typescript-eslint/eslint-plugin": "4.20.0",
+    "@typescript-eslint/parser": "4.20.0",
     "discord-api-types": "0.13.3",
     "eslint": "7.23.0",
     "husky": "6.0.0",
