+++ src/automod/Shortlinks.ts
@@ -21,17 +21,21 @@
  */
 
 import type { Message, TextChannel } from 'eris';
+import { Inject, LinkParent } from '@augu/lilith';
 import PunishmentService from '../services/PunishmentService';
 import * as Constants from '../util/Constants';
 import PermissionUtil from '../util/Permissions';
+import AutomodService from '../services/AutomodService';
 import { Automod } from '../structures';
-import { Inject } from '@augu/lilith';
 import Database from '../components/Database';
 import Discord from '../components/Discord';
 
+@LinkParent(AutomodService)
 export default class Shortlinks implements Automod {
-  private punishments!: PunishmentService;
   public name: string = 'shortlinks';
+
+  @Inject
+  private punishments!: PunishmentService;
 
   @Inject
   private database!: Database;
