+++ src/commands/core/HelpCommand.ts
@@ -22,11 +22,11 @@
 
 import { Command, CommandMessage, EmbedBuilder } from '../../structures';
 import { Constants as ErisConstants } from 'eris';
+import { Inject, LinkParent } from '@augu/lilith';
 import { firstUpper } from '@augu/utils';
 import * as Constants from '../../util/Constants';
 import CommandService from '../../services/CommandService';
 import Permissions from '../../util/Permissions';
-import { Inject } from '@augu/lilith';
 
 interface CommandCategories {
   moderation?: Command[];
@@ -34,6 +34,7 @@ interface CommandCategories {
   general?: Command[];
 }
 
+@LinkParent(CommandService)
 export default class HelpCommand extends Command {
   private categories!: CommandCategories;
 
@@ -65,7 +66,7 @@ export default class HelpCommand extends Command {
     if (this.categories === undefined) {
       this.categories = {};
 
-      const commands = this.service.commands.filter(cmd => !cmd.ownerOnly);
+      const commands = this.service.filter(cmd => !cmd.ownerOnly);
       for (let i = 0; i < commands.length; i++) {
         const command = commands[i];
         (this.categories[command.category] ??= []).push(command);
@@ -79,7 +80,7 @@ export default class HelpCommand extends Command {
         `:pencil2: **For more documentation of a command or module, run \`${prefix}help <cmdOrMod>\` with \`<cmdOrMod>\` with the command or module you want to look up.**`,
         '',
         'You can browse the [website](https://nino.floofy.dev) for more information and a prettier UI for this help command.',
-        `There are currently **${this.service.commands.size}** commands available.`
+        `There are currently **${this.service.size}** commands available.`
       ]);
 
     for (const cat in (this.categories as Required<CommandCategories>)) {
@@ -91,7 +92,7 @@ export default class HelpCommand extends Command {
   }
 
   private async renderDoc(msg: CommandMessage, cmdOrMod: string) {
-    const command = this.service.commands.filter(cmd => !cmd.hidden && cmd.name === cmdOrMod || cmd.aliases.includes(cmdOrMod))[0];
+    const command = this.service.filter(cmd => !cmd.hidden && cmd.name === cmdOrMod || cmd.aliases.includes(cmdOrMod))[0];
     const prefix = msg.settings.prefixes[msg.settings.prefixes.length - 1];
 
     if (command !== undefined) {
@@ -128,12 +129,12 @@ export default class HelpCommand extends Command {
             },
             {
               name: '• User Permissions',
-              value: Permissions.stringify(command.userPermissions.reduce((acc, curr) => acc | ErisConstants.Permissions[curr], 0)) || 'None',
+              value: Permissions.stringify(command.userPermissions.reduce((acc, curr) => acc | ErisConstants.Permissions[curr], 0n)) || 'None',
               inline: true
             },
             {
               name: '• Bot Permissions',
-              value: Permissions.stringify(command.botPermissions.reduce((acc, curr) => acc | ErisConstants.Permissions[curr], 0)) || 'None',
+              value: Permissions.stringify(command.botPermissions.reduce((acc, curr) => acc | ErisConstants.Permissions[curr], 0n)) || 'None',
               inline: true
             },
             {
@@ -146,7 +147,7 @@ export default class HelpCommand extends Command {
 
       return msg.reply(embed);
     } else {
-      const mod = this.service.commands.filter(cmd => cmd.category.toLowerCase() === cmdOrMod.toLowerCase());
+      const mod = this.service.filter(cmd => cmd.category.toLowerCase() === cmdOrMod.toLowerCase());
       if (mod.length > 0) {
         const longestName = this._calculateLength(mod.sort((a, b) => this._calculateLength(b) - this._calculateLength(a))[0]);
         const embed = new EmbedBuilder()
