+++ src/commands/core/InviteCommand.ts
@@ -21,9 +21,11 @@
  */
 
 import { Command, CommandMessage } from '../../structures';
-import { Inject } from '@augu/lilith';
+import { Inject, LinkParent } from '@augu/lilith';
+import CommandService from '../../services/CommandService';
 import Discord from '../../components/Discord';
 
+@LinkParent(CommandService)
 export default class InviteCommand extends Command {
   @Inject
   private discord!: Discord;
