+++ src/commands/moderation/BanCommand.ts
@@ -22,10 +22,11 @@
 
 import { Command, CommandMessage } from '../../structures';
 import { DiscordRESTError, User } from 'eris';
+import { Inject, LinkParent } from '@augu/lilith';
 import { PunishmentType } from '../../entities/PunishmentsEntity';
 import PunishmentService from '../../services/PunishmentService';
+import CommandService from '../../services/CommandService';
 import { Categories } from '../../util/Constants';
-import { Inject } from '@augu/lilith';
 import Discord from '../../components/Discord';
 import ms = require('ms');
 
@@ -35,6 +36,7 @@ interface BanFlags {
   d?: string | true;
 }
 
+@LinkParent(CommandService)
 export default class BanCommand extends Command {
   @Inject
   private punishments!: PunishmentService;
