+++ src/listeners/VoiceStateListener.ts
@@ -21,12 +21,14 @@
  */
 
 import type { Member, TextChannel, VoiceChannel } from 'eris';
+import { Inject, LinkParent } from '@augu/lilith';
 import { LoggingEvents } from '../entities/LoggingEntity';
-import { Inject } from '@augu/lilith';
+import ListenerService from '../services/ListenerService';
 import Subscribe from '../structures/decorators/Subscribe';
 import Database from '../components/Database';
 import Discord from '../components/Discord';
 
+@LinkParent(ListenerService)
 export default class VoiceStateListener {
   @Inject
   private database!: Database;
