+++ src/services/CommandService.ts
@@ -20,24 +20,34 @@
  * SOFTWARE.
  */
 
-import { Service, Inject, getInjectables } from '@augu/lilith';
-import type { Message, TextChannel } from 'eris';
-import { readdirSync, Ctor } from '@augu/utils';
+import { Service, Inject, FindChildrenIn } from '@augu/lilith';
+import { EmbedBuilder, CommandMessage } from '../structures';
+import type { Message, OldMessage, TextChannel } from 'eris';
+import LocalizationService from './LocalizationService';
 import type NinoCommand from '../structures/Command';
-import CommandHandler from '../structures/handlers/CommandHandler';
 import { Collection } from '@augu/collections';
+import Subcommand from '../structures/Subcommand';
+import Prometheus from '../components/Prometheus';
 import { Logger } from 'tslog';
+import Database from '../components/Database';
 import { join } from 'path';
 import Discord from '../components/Discord';
-import app from '../container';
+import Config from '../components/Config';
 
-export default class CommandService implements Service {
+const FLAG_REGEX = /(?:--?|—)([\w]+)(=?(\w+|['"].*['"]))?/gi;
+
+@Service({
+  priority: 0,
+  name: 'commands'
+})
+@FindChildrenIn(join(process.cwd(), 'commands'))
+export default class CommandService extends Collection<string, NinoCommand> {
   public commandsExecuted: number = 0;
   public messagesSeen: number = 0;
   public cooldowns: Collection<string, Collection<string, number>> = new Collection();
-  public commands: Collection<string, NinoCommand> = new Collection();
-  private handler: CommandHandler = new CommandHandler(this);
-  public name = 'Commands';
+
+  @Inject
+  private config!: Config;
 
   @Inject
   private logger!: Logger;
@@ -45,47 +55,193 @@ export default class CommandService implements Service {
   @Inject
   private discord!: Discord;
 
-  async load() {
-    // Add injections to the command handler
-    const injections = getInjectables(this.handler);
-    app.inject(injections, this.handler);
+  @Inject
+  private database!: Database;
 
-    // Add in the message create and event
-    this.discord.client.on('messageCreate', (msg: Message<TextChannel>) =>
-      this.handler.handleCommand.call(this.handler, msg)
+  @Inject
+  private prometheus!: Prometheus;
+
+  @Inject
+  private localization!: LocalizationService;
+
+  onChildLoad(command: NinoCommand) {
+    this.logger.info(`✔ Loaded command ${command.name}!`);
+    this.set(command.name, command);
+  }
+
+  async handleCommand(msg: Message<TextChannel>) {
+    this.prometheus.messagesSeen?.inc();
+    this.messagesSeen++;
+
+    if (msg.author.bot) return;
+    if (![0, 5].includes(msg.channel.type)) return;
+
+    const settings = await this.database.guilds.get(msg.channel.guild.id);
+    const userSettings = await this.database.users.get(msg.author.id);
+
+    // remove duplicates
+    const prefixes = [...new Set([`${this.discord.mentionRegex}`].concat(settings.prefixes, userSettings.prefixes, this.config.getProperty('prefixes')!).filter(Boolean))];
+    if (this.discord.client.user.id === '531613242473054229')
+      prefixes.push('nino ');
+
+    const prefix = prefixes.find(prefix => msg.content.startsWith(prefix));
+    if (prefix === undefined)
+      return;
+
+    let rawArgs = msg.content.slice(prefix.length).trim().split(/ +/g);
+    const name = rawArgs.shift()!;
+    const command = this.find(command =>
+      command.name === name || command.aliases.includes(name)
     );
 
-    this.discord.client.on('messageUpdate', (msg: Message<TextChannel>, old) =>
-      this.handler.onMessageEdit.call(this.handler, msg, old)
-    );
+    if (command === null)
+      return;
 
-    // Load in commands
-    this.logger.info('Now loading in commands...');
-    const files = readdirSync(join(__dirname, '..', 'commands'));
+    // Check for if the guild is blacklisted
+    const guildBlacklist = await this.database.blacklists.get(msg.guildID);
+    if (guildBlacklist !== undefined) {
+      const issuer = this.discord.client.users.get(guildBlacklist.issuer);
+      await msg.channel.createMessage([
+        `:pencil2: **This guild is blacklisted by ${issuer?.username ?? 'Unknown User'}#${issuer?.discriminator ?? '0000'}**`,
+        `> ${guildBlacklist.reason ?? '*(no reason provided)*'}`,
+        '',
+        'If there is a issue or want to be unblacklisted, reach out to the developers here: discord.gg/ATmjFH9kMH in under #support.',
+        'I will attempt to leave this guild, goodbye. :wave:'
+      ].join('\n'));
 
-    for (let i = 0; i < files.length; i++) {
-      const file = files[i];
-      const ctor: Ctor<any> = await import(file);
-      const command: NinoCommand = new ctor.default!();
-
-      // only the help command will have the `commands` prop to be this
-      const injections = getInjectables(command);
-      app.inject(injections.filter(inject => inject.property !== 'service'), command);
-
-      // if it's the help command, let's inject it
-      const injection = injections.filter(inject => inject.property === 'service')[0];
-      if (injection !== undefined)
-        app.inject([
-          {
-            property: 'service',
-            ref: CommandService
-          }
-        ], command);
-
-      this.commands.set(command.name, command);
-      this.logger.info(`Initialized command ${command.name}`);
+      await msg.channel.guild.leave();
+      return;
     }
 
-    this.logger.info(`Loaded ${this.commands.size} commands!`);
+    // Check if the user is blacklisted
+    const userBlacklist = await this.database.blacklists.get(msg.author.id);
+    if (userBlacklist !== undefined) {
+      const issuer = this.discord.client.users.get(userBlacklist.issuer);
+      return msg.channel.createMessage([
+        `:pencil2: **You were blacklisted by ${issuer?.username ?? 'Unknown User'}#${issuer?.discriminator ?? '0000'}**`,
+        `> ${userBlacklist.reason ?? '*(no reason provided)*'}`,
+        '',
+        'If there is a issue or want to be unblacklisted, reach out to the developers here: discord.gg/ATmjFH9kMH in under #support.'
+      ].join('\n'));
+    }
+
+    const locale = this.localization.get(settings.language, userSettings.language);
+    const message = new CommandMessage(msg, locale, settings, userSettings);
+    const owners = this.config.getProperty('owners') ?? [];
+    if (command.ownerOnly && !owners.includes(msg.author.id))
+      return message.reply(`Command **${command.name}** is a developer-only command, nice try...`);
+
+    // Check for permissions of Nino
+    if (command.botPermissions.length) {
+      const permissions = msg.channel.permissionsOf(this.discord.client.user.id);
+      const missing = command.botPermissions.filter(perm => !permissions.has(perm));
+
+      if (missing.length > 0)
+        return message.reply(`I am missing the following permissions: **${missing.join(', ')}**`);
+    }
+
+    // Check for the user's permissions
+    if (command.userPermissions.length) {
+      const permissions = msg.channel.permissionsOf(msg.author.id);
+      const missing = command.userPermissions.filter(perm => !permissions.has(perm));
+
+      if (missing.length > 0)
+        return message.reply(`You are missing the following permission: **${missing.join(', ')}**`);
+    }
+
+    // Cooldowns
+    if (!this.cooldowns.has(command.name))
+      this.cooldowns.set(command.name, new Collection());
+
+    const now = Date.now();
+    const timestamps = this.cooldowns.get(command.name)!;
+    const amount = command.cooldown * 1000;
+
+    if (!owners.includes(msg.author.id) && timestamps.has(msg.author.id)) {
+      const time = timestamps.get(msg.author.id)! + amount;
+      if (now < time) {
+        const left = (time - now) / 1000;
+        return message.reply(`Please wait **${left.toFixed()}** seconds before executing this command.`);
+      }
+    }
+
+    timestamps.set(msg.author.id, now);
+    setTimeout(() => timestamps.delete(msg.author.id), amount);
+
+    // Figure out the subcommand
+    let methodName = 'run';
+    let subcommand: Subcommand | undefined = undefined;
+    for (const arg of rawArgs) {
+      if (command.subcommands.length && command.subcommands.find(r => r.name === arg) !== undefined) {
+        subcommand = command.subcommands.find(r => r.name === arg)!;
+        methodName = subcommand.name;
+
+        break;
+      }
+    }
+
+    if (subcommand !== undefined)
+      rawArgs.shift();
+
+    message['_flags'] = this.parseFlags(rawArgs.join(' '));
+
+    // Sanatize command args
+    if (command.name !== 'eval') {
+      rawArgs = rawArgs.filter(arg => !FLAG_REGEX.test(arg));
+    }
+
+    try {
+      const executor = Reflect.get(command, methodName);
+      if (typeof executor !== 'function')
+        throw new SyntaxError(`${subcommand ? 'Subc' : 'C'}ommand "${subcommand ? methodName : command.name}" was not a function.`);
+
+      this.prometheus.commandsExecuted?.labels(command.name).inc();
+      this.commandsExecuted++;
+      await executor.call(command, message, rawArgs);
+      this.logger.info(`Command "${command.name}" has been ran by ${msg.author.username}#${msg.author.discriminator} in guild ${msg.channel.guild.name} (${msg.channel.guild.id})`);
+    } catch(ex) {
+      const _owners = await Promise.all(owners.map(id => {
+        const user = this.discord.client.users.get(id);
+        if (user === undefined)
+          return this.discord.client.getRESTUser(id);
+        else
+          return Promise.resolve(user);
+      }));
+
+      const contact = _owners
+        .map((r, index) => `${index + 1 === owners.length ? 'or ' : ''}**${r.username}#${r.discriminator}**`)
+        .join(', ');
+
+      const embed = new EmbedBuilder()
+        .setColor(0xDAA2C6)
+        .setDescription([
+          `${subcommand !== undefined ? `Subcommand **${methodName}** (parent **${command.name}**)` : `Command **${command.name}**`} has failed to execute.`,
+          `If this is a re-occuring issue, contact ${contact} at <https://discord.gg/JjHGR6vhcG>, under the <#747522228714733610> channel.`,
+          '',
+          '```js',
+          ex.stack ?? '<... no stacktrace? ...>',
+          '```'
+        ])
+        .build();
+
+      await msg.channel.createMessage({ embed });
+
+      // hacky way for "TypeError  Cannot set property name of  which has only a getter"
+      this.logger.error(`${subcommand !== undefined ? `Subcommand ${methodName}` : `Command ${command.name}`} has failed to execute`);
+      console.error(ex.stack);
+    }
+  }
+
+  // credit for regex: Ice <3
+  private parseFlags(content: string): Record<string, string | true> {
+    const record: Record<string, string | true> = {};
+    content.replaceAll(FLAG_REGEX, (_, key: string, value: string) => {
+      record[key.trim()] = value ? value.replaceAll(/(^[='"]+|['"]+$)/g, '').trim() : true;
+      return value;
+    });
+
+    // keep it immutable so
+    // the application doesn't mutate its state
+    return Object.freeze(record);
   }
 }
